+++ src/components/blocks/order/OrderItemMaterials.svelte
@@ -8,7 +8,7 @@
   import TextInput from "./common/TextInput.svelte";
   import { slide } from "svelte/transition";
   import type { IProduct } from "@/lib/types.svelte";
-  import { resolveColorCount } from "@/lib/materialUtils";
+  import { resolveColorCount } from "@/lib/pricing/materials";
   import type { ProductMaterialValue } from "@/lib/types.svelte";
 
   interface Props {
