+++ src/lib/__tests__/fieldValue.test.ts
@@ -5,7 +5,7 @@
  * coercing a raw string. These tests lock that contract in one place.
  */
 import { describe, expect, it } from "vitest";
-import { findFieldByName, resolveFieldValue, resolveNumericValue } from "@/lib/fieldValue";
+import { findFieldByName, resolveFieldValue, resolveNumericValue } from "@/lib/pricing/fieldValue";
 import { makeField } from "./fixtures";
 
 describe("resolveFieldValue", () => {
